requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.3.3
openpyxl==3.1.5
openai==1.12.0
python-dotenv==1.0.0
Pillow==10.2.0
slowapi==0.1.9

# SQLAdmin для админ-панели
sqladmin==0.17.0
sqlalchemy==2.0.23

# PostgreSQL driver
psycopg2-binary==2.9.9

# Auth
PyJWT>=2.8.0
google-auth>=2.27.0


